telux: Sensor test app improvements
Sensor test app improvements
Remove unused forward declaration
Added details to sensor API behavior

CRs-Fixed: 2984229

+++ include/telux/sensor/SensorDefines.hpp
@@ -107,7 +107,7 @@ struct SensorInfo {
      * For gyroscopes, this is the number of degrees per second (dps) in either direction (+/-)
      * along each axis
      */
-    float range;
+    int range;
 };
 
 /**
